Refactor Threading for Platform Safety & Update Bolt Roadmap
- Refactored `AestraCore` threading: moved `MMCSS`, `ThreadPool`, `RealTimeThreadPool` implementations to `src/AestraThreading.cpp` to isolate platform dependencies.
- Removed platform leaks: cleaned `<windows.h>` from `AestraThreading.h` and `AudioEngine.h`.
- Added `// ALLOW_PLATFORM_INCLUDE` to `ASIOInterface.h`.
- Updated `scripts/audit_codebase.py` to ignore comments (fixing `EffectChain` false positive).
- Updated `bolt.md` with new innovations: NeuralMix Assistant, Collaborative Editing, GPU-DSP, Voice Cloning.
- Fixed missing `<cstring>` include in `AestraFile.h`.
