+++ AestraAudio/include/Core/AudioEngine.h
@@ -16,10 +16,6 @@
 #include <vector>
 #include <atomic>
 
-#ifdef _WIN32
-#define WIN32_LEAN_AND_MEAN
-#include <windows.h>
-#endif
 #include <memory>
 #include <mutex>
 #include <array>
